refactor: projection dto 패키지 이동
