+++ app-main/src/main/java/net/causw/app/main/domain/community/report/api/v1/mapper/ReportDtoMapper.java
@@ -6,9 +6,9 @@ import org.mapstruct.Named;
 import org.mapstruct.factory.Mappers;
 
 import net.causw.app.main.domain.community.report.api.v1.dto.ReportCreateResponseDto;
-import net.causw.app.main.domain.community.report.api.v1.dto.ReportedCommentNativeProjection;
+import net.causw.app.main.domain.community.report.repository.projection.ReportedCommentNativeProjection;
 import net.causw.app.main.domain.community.report.api.v1.dto.ReportedCommentResponseDto;
-import net.causw.app.main.domain.community.report.api.v1.dto.ReportedPostNativeProjection;
+import net.causw.app.main.domain.community.report.repository.projection.ReportedPostNativeProjection;
 import net.causw.app.main.domain.community.report.api.v1.dto.ReportedPostResponseDto;
 import net.causw.app.main.domain.community.report.api.v1.dto.ReportedUserResponseDto;
 import net.causw.app.main.domain.community.report.enums.ReportReason;
