+++ app-main/src/main/java/net/causw/app/main/domain/community/report/repository/projection/ReportedCommentNativeProjection.java
@@ -1,4 +1,4 @@
-package net.causw.app.main.domain.community.report.api.v1.dto;
+package net.causw.app.main.domain.community.report.repository.projection;
 
 import java.time.LocalDateTime;
 
@@ -6,14 +6,20 @@ import net.causw.app.main.domain.user.account.enums.user.UserState;
 
 /**
  * Native Query 결과를 받기 위한 Projection 인터페이스
- * 게시글 신고를 조회할 때 사용
+ * 댓글과 대댓글 신고를 UNION ALL로 합쳐서 조회할 때 사용
  */
-public interface ReportedPostNativeProjection {
+public interface ReportedCommentNativeProjection {
 	String getReportId();
+
+	String getContentId();
+
+	String getContent();
+
+	String getPostTitle();
 
 	String getPostId();
 
-	String getPostTitle();
+	String getBoardId();
 
 	String getWriterName();
 
@@ -22,8 +28,4 @@ public interface ReportedPostNativeProjection {
 	String getReportReason();
 
 	LocalDateTime getReportCreatedAt();
-
-	String getBoardName();
-
-	String getBoardId();
 }
