+++ app-main/src/main/java/net/causw/app/main/domain/community/report/repository/projection/ReportedPostNativeProjection.java
@@ -1,4 +1,4 @@
-package net.causw.app.main.domain.community.report.api.v1.dto;
+package net.causw.app.main.domain.community.report.repository.projection;
 
 import java.time.LocalDateTime;
 
@@ -6,20 +6,14 @@ import net.causw.app.main.domain.user.account.enums.user.UserState;
 
 /**
  * Native Query 결과를 받기 위한 Projection 인터페이스
- * 댓글과 대댓글 신고를 UNION ALL로 합쳐서 조회할 때 사용
+ * 게시글 신고를 조회할 때 사용
  */
-public interface ReportedCommentNativeProjection {
+public interface ReportedPostNativeProjection {
 	String getReportId();
-
-	String getContentId();
-
-	String getContent();
-
-	String getPostTitle();
 
 	String getPostId();
 
-	String getBoardId();
+	String getPostTitle();
 
 	String getWriterName();
 
@@ -28,4 +22,8 @@ public interface ReportedCommentNativeProjection {
 	String getReportReason();
 
 	LocalDateTime getReportCreatedAt();
+
+	String getBoardName();
+
+	String getBoardId();
 }
