+++ app-main/src/main/java/net/causw/app/main/domain/community/report/service/v2/dto/ReportedPostSummaryResult.java
@@ -2,7 +2,7 @@ package net.causw.app.main.domain.community.report.service.v2.dto;
 
 import java.time.LocalDateTime;
 
-import net.causw.app.main.domain.community.report.api.v1.dto.ReportedPostNativeProjection;
+import net.causw.app.main.domain.community.report.repository.projection.ReportedPostNativeProjection;
 import net.causw.app.main.domain.community.report.enums.ReportReason;
 import net.causw.app.main.domain.user.account.enums.user.UserState;
