+++ app-main/src/main/java/net/causw/app/main/domain/community/report/service/v2/implementation/PostReportReader.java
@@ -5,7 +5,7 @@ import org.springframework.transaction.annotation.Transactional;
 import org.springframework.data.domain.Page;
 import org.springframework.data.domain.Pageable;
 
-import net.causw.app.main.domain.community.report.api.v1.dto.ReportedPostNativeProjection;
+import net.causw.app.main.domain.community.report.repository.projection.ReportedPostNativeProjection;
 import net.causw.app.main.domain.community.report.enums.ReportType;
 import net.causw.app.main.domain.community.report.repository.ReportRepository;
 import net.causw.app.main.domain.user.account.entity.user.User;
