+++ app-main/src/test/java/net/causw/app/main/domain/community/report/service/v1/ReportServiceTest.java
@@ -29,9 +29,9 @@ import net.causw.app.main.domain.community.post.entity.Post;
 import net.causw.app.main.domain.community.post.repository.PostRepository;
 import net.causw.app.main.domain.community.report.api.v1.dto.ReportCreateRequestDto;
 import net.causw.app.main.domain.community.report.api.v1.dto.ReportCreateResponseDto;
-import net.causw.app.main.domain.community.report.api.v1.dto.ReportedCommentNativeProjection;
+import net.causw.app.main.domain.community.report.repository.projection.ReportedCommentNativeProjection;
 import net.causw.app.main.domain.community.report.api.v1.dto.ReportedCommentResponseDto;
-import net.causw.app.main.domain.community.report.api.v1.dto.ReportedPostNativeProjection;
+import net.causw.app.main.domain.community.report.repository.projection.ReportedPostNativeProjection;
 import net.causw.app.main.domain.community.report.api.v1.dto.ReportedPostResponseDto;
 import net.causw.app.main.domain.community.report.api.v1.dto.ReportedUserResponseDto;
 import net.causw.app.main.domain.community.report.api.v1.mapper.ReportDtoMapper;
